refactor navigation: move view mode controls below fretboard
Simplifies EnhancedNavigationPanel to focus on chord selection only, and creates dedicated PositionControls component that handles both position navigation and view mode toggles. Controls are now positioned contextually near the fretboard they control.
